+++ netutil/netutil_linux.go
@@ -1,4 +1,3 @@
-// Package netutil provides methods for working with network
 package netutil
 
 // ////////////////////////////////////////////////////////////////////////////////// //
@@ -10,9 +9,7 @@ package netutil
 
 import (
 	"bufio"
-	"net"
 	"os"
-	"strings"
 
 	"pkg.re/essentialkaos/ek.v12/strutil"
 )
@@ -23,55 +20,6 @@ import (
 var procRouteFile = "/proc/net/route"
 
 // ////////////////////////////////////////////////////////////////////////////////// //
-
-// GetIP returns main IPv4 address
-func GetIP() string {
-	return getMainIP(false)
-}
-
-// GetIP6 returns main IPv6 address
-func GetIP6() string {
-	return getMainIP(true)
-}
-
-// ////////////////////////////////////////////////////////////////////////////////// //
-
-func getMainIP(v6 bool) string {
-	interfaces, err := net.Interfaces()
-
-	if err != nil {
-		return ""
-	}
-
-	defaultInterface := getDefaultRouteInterface()
-
-	for i := len(interfaces) - 1; i >= 0; i-- {
-		// Ignore TUN/TAP interfaces
-		if strings.HasPrefix(interfaces[i].Name, "t") {
-			continue
-		}
-
-		if defaultInterface != "" && interfaces[i].Name != defaultInterface {
-			continue
-		}
-
-		addr, err := interfaces[i].Addrs()
-
-		if err != nil || len(addr) == 0 {
-			continue
-		}
-
-		for _, a := range addr {
-			ipnet, ok := a.(*net.IPNet)
-
-			if ok && strings.Contains(ipnet.IP.String(), "::") == v6 {
-				return ipnet.IP.String()
-			}
-		}
-	}
-
-	return ""
-}
 
 func getDefaultRouteInterface() string {
 	fd, err := os.OpenFile(procRouteFile, os.O_RDONLY, 0)